requirements.txt
# requirements.txt — Omnira Synora
fastapi==0.115.0
uvicorn[standard]==0.30.0
pillow==10.4.0
numpy==1.26.4
pydantic==2.8.0
pydantic-settings==2.3.0
sqlalchemy==2.0.31
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
httpx==0.27.0
python-multipart==0.0.9

# Opsional — aktifkan sesuai storage yang dipakai:
# supabase==2.5.0      # untuk Supabase storage
# boto3==1.34.0        # untuk AWS S3
# asyncpg==0.29.0      # untuk PostgreSQL production
